Ensure typings files are type-checked despite skipLibCheck
Change-type: patch
