+++ package.json
@@ -26,8 +26,8 @@
     "build": "npm run clean && tsc --project ./tsconfig.build.json && copyup \"src/**/*.sbvr\" \"src/**/*.sql\" dist/",
     "clean": "rimraf dist/",
     "cleanup": "rm .fast-boot.json .pinejs-cache.json || true",
-    "lint": "balena-lint src/ test/ typings/ init.ts index.js && tsc --noEmit --project .",
-    "prettify": "balena-lint --fix src/ test/ typings/ init.ts index.js",
+    "lint": "balena-lint src/ test/ init.ts index.js && tsc --noEmit --project .",
+    "prettify": "balena-lint --fix src/ test/ init.ts index.js",
     "prepack": "npm run build",
     "test:compose": "./automation/test.sh",
     "test": "npm run lint",
