+++ src/typings/thirty-two.ts
@@ -1,3 +1,5 @@
+import './thirty-two-shim';
+
 declare module 'thirty-two' {
 	export const encode: (key: string | Buffer) => string;
 }
