--- typings/memoizee.d.ts
@@ -1,4 +1,0 @@
-declare module 'memoizee/normalizers/primitive' {
-	function PrimitiveNormalizer(args: any[]): string;
-	export = PrimitiveNormalizer;
-}
